Refactor property names & add comments
